feat(analytics): move Expiring/Deployed Today and Workhorse panels from /home to /analytics' Apps tab
