+++ client/src/home/HomeOverview/index.jsx
@@ -12,9 +12,6 @@ import ReactCountryFlag from 'react-country-flag';
 import CountUp from 'components/CountUp';
 
 import { CC_COLLATERAL_CUMULUS, CC_COLLATERAL_NIMBUS, CC_COLLATERAL_STRATUS } from 'content';
-import { APP_CATEGORY_META } from 'content/appCategoryMeta';
-import { CategoryTooltip } from 'components/CategoryTooltip';
-import { WorkhorsePanel } from 'home/WorkhorsePanel';
 import { AppEcosystemBreakdown } from 'components/AppEcosystemBreakdown';
 import { TopHostedApps } from 'components/TopHostedApps';
 import { fluxos_version_string, daemon_version_string } from 'main/flux_version';
@@ -31,14 +28,6 @@ function fmtCompact(n) {
   if (n >= 1000000) return (n / 1000000).toFixed(1) + 'M';
   if (n >= 1000) return (n / 1000).toFixed(1) + 'k';
   return n.toFixed(0);
-}
-
-function blocksToHuman(blocks) {
-  const totalMinutes = Math.round(blocks * 0.5);
-  if (totalMinutes < 60) return `${totalMinutes}m`;
-  const h = Math.floor(totalMinutes / 60);
-  const m = totalMinutes % 60;
-  return m > 0 ? `${h}h ${m}m` : `${h}h`;
 }
 
 // ── Shared sub-components ──────────────────────────────────────────────────────
@@ -299,122 +288,6 @@ function NetworkResourcesPanel({ gstore }) {
   );
 }
 
-// ── Spec Header (shared by Expiring / Deployed panels) ──────────────────────
-
-/*
- * Enterprise apps ship an encrypted compose, so their CPU / RAM / SSD are
- * genuinely unknown — not zero. Render an em dash rather than a confident 0.00.
- */
-function fmtSpecVal(value, suffix) {
-  if (value == null) return '—';
-  return `${value.toFixed(2)}${suffix}`;
-}
-
-
-function SpecHeader() {
-  return (
-    <div className="hov-spec-header">
-      <span className="hov-spec-header__name">Name</span>
-      <span className="hov-spec-header__cat">Cat</span>
-      <span className="hov-spec-header__inst">Inst</span>
-      <span className="hov-spec-header__val">CPU</span>
-      <span className="hov-spec-header__val">RAM</span>
-      <span className="hov-spec-header__val">SSD</span>
-      <span className="hov-spec-header__time">Time</span>
-    </div>
-  );
-}
-
-function SpecCategoryIcon({ category }) {
-  const meta = APP_CATEGORY_META[category] || APP_CATEGORY_META.other;
-  const { Icon, color } = meta;
-  // Per-app row: the network-wide breakdown would be misleading here.
-  const tooltip = <CategoryTooltip category={category} />;
-  return (
-    <Tooltip2 content={tooltip} placement="top" hoverOpenDelay={200} popoverClassName="hov-cat-tooltip">
-      <span className="hov-spec-cat" style={{ color }}>
-        <Icon size={11} />
-      </span>
-    </Tooltip2>
-  );
-}
-
-// ── Panel 5: Apps Expiring Today ──────────────────────────────────────────────
-
-function ExpiringTodayPanel({ appSpecs }) {
-  if (!appSpecs) {
-    return (
-      <div className="hov-panel hov-panel-center">
-        <Spinner size={24} />
-      </div>
-    );
-  }
-
-  const items = appSpecs.expiringToday || [];
-
-  return (
-    <div className="hov-panel hov-panel--expiring">
-      <PanelHeader title="EXPIRING TODAY" badge={items.length || null} />
-      {items.length > 0 && <SpecHeader />}
-      <div className="hov-list">
-        {items.length === 0 ? (
-          <div className="hov-empty">None expiring today</div>
-        ) : (
-          items.map((spec, i) => (
-            <div key={spec.name + i} className="hov-spec-row">
-              <span className="hov-list-name">{spec.name}</span>
-              <SpecCategoryIcon category={spec.category} />
-              <span className="hov-badge hov-badge--warn">{spec.instances}×</span>
-              <span className="hov-spec-val">{fmtSpecVal(spec.cpuPerInst, 'c')}</span>
-              <span className="hov-spec-val">{fmtSpecVal(spec.ramGBPerInst, 'GB')}</span>
-              <span className="hov-spec-val">{fmtSpecVal(spec.ssdGBPerInst, 'GB')}</span>
-              <span className="hov-time hov-time--warn">in {blocksToHuman(spec.expiresInBlocks)}</span>
-            </div>
-          ))
-        )}
-      </div>
-    </div>
-  );
-}
-
-// ── Panel 6: Deployed Today ───────────────────────────────────────────────────
-
-function DeployedTodayPanel({ appSpecs }) {
-  if (!appSpecs) {
-    return (
-      <div className="hov-panel hov-panel-center">
-        <Spinner size={24} />
-      </div>
-    );
-  }
-
-  const items = appSpecs.deployedToday || [];
-
-  return (
-    <div className="hov-panel hov-panel--deployed">
-      <PanelHeader title="DEPLOYED TODAY" badge={items.length || null} />
-      {items.length > 0 && <SpecHeader />}
-      <div className="hov-list">
-        {items.length === 0 ? (
-          <div className="hov-empty">None deployed today</div>
-        ) : (
-          items.map((spec, i) => (
-            <div key={spec.name + i} className="hov-spec-row">
-              <span className="hov-list-name">{spec.name}</span>
-              <SpecCategoryIcon category={spec.category} />
-              <span className="hov-badge hov-badge--green">{spec.instances}×</span>
-              <span className="hov-spec-val">{fmtSpecVal(spec.cpuPerInst, 'c')}</span>
-              <span className="hov-spec-val">{fmtSpecVal(spec.ramGBPerInst, 'GB')}</span>
-              <span className="hov-spec-val">{fmtSpecVal(spec.ssdGBPerInst, 'GB')}</span>
-              <span className="hov-time hov-time--green">{blocksToHuman(spec.deployedAgeBlocks)} ago</span>
-            </div>
-          ))
-        )}
-      </div>
-    </div>
-  );
-}
-
 // ── Panel 7: Node Geolocation ─────────────────────────────────────────────────
 
 function GeoDistributionPanel({ gstore, countryCounts }) {
@@ -565,7 +438,7 @@ function FluxAIPanel({ gpuPrices }) {
 
 // ── Main export ───────────────────────────────────────────────────────────────
 
-export function HomeOverview({ gstore, appSpecs, countryCounts, gpuPrices }) {
+export function HomeOverview({ gstore, countryCounts, gpuPrices }) {
   return (
     <div className="home-overview">
       <div className="home-overview-row">
@@ -573,12 +446,7 @@ export function HomeOverview({ gstore, appSpecs, countryCounts, gpuPrices }) {
         <NetworkResourcesPanel gstore={gstore} />
         <AppEcosystemBreakdown gstore={gstore} />
       </div>
-      <div className="home-overview-row home-overview-row--bottom">
-        <TopHostedApps gstore={gstore} />
-        <ExpiringTodayPanel appSpecs={appSpecs} />
-        <DeployedTodayPanel appSpecs={appSpecs} />
-      </div>
-      <WorkhorsePanel gstore={gstore} appSpecs={appSpecs} />
+      <TopHostedApps gstore={gstore} />
       {SHOW_FLUX_AI_PANEL && <FluxAIPanel gpuPrices={gpuPrices} />}
       <GeoDistributionPanel gstore={gstore} countryCounts={countryCounts} />
     </div>
